+++ src/sbm_model.jl
@@ -28,12 +28,6 @@ function initialize_sbm_model(config::Config)
     nc = NCDataset(static_path)
     dims = dimnames(nc[param(config, "input.subcatchment")])
 
-    # There is no need to permute the dimensions of the data, since the active indices are
-    # correctly calculated in both ways.
-    # The dimension order only needs to be known for interpreting the LDD directions
-    # and creating the coordinate maps.
-    dims_xy = dims[2] in ("y", "lat")
-
     subcatch_2d = ncread(nc, param(config, "input.subcatchment"); allow_missing = true)
     # indices based on catchment
     inds, rev_inds = active_indices(subcatch_2d, missing)
@@ -50,13 +44,9 @@ function initialize_sbm_model(config::Config)
     riverlength = riverlength_2d[inds]
 
     # read x, y coordinates and calculate cell length [m]
-    y_nc = "y" in keys(nc.dim) ? ncread(nc, "y") : ncread(nc, "lat")
-    x_nc = "x" in keys(nc.dim) ? ncread(nc, "x") : ncread(nc, "lon")
-    if dims_xy
-        y = permutedims(repeat(y_nc, outer = (1, length(x_nc))))[inds]
-    else
-        y = repeat(y_nc, outer = (1, length(x_nc)))[inds]
-    end
+    y_nc = read_y_axis(nc)
+    x_nc = read_x_axis(nc)
+    y = permutedims(repeat(y_nc, outer = (1, length(x_nc))))[inds]
     cellength = abs(mean(diff(x_nc)))
 
     xl = fill(mv, n)
@@ -197,7 +187,6 @@ function initialize_sbm_model(config::Config)
         lake = nothing,
     )
 
-    pcr_dir = dims_xy ? permute_indices(pcrdir) : pcrdir
     graph = flowgraph(ldd, inds, pcr_dir)
 
 
@@ -267,7 +256,6 @@ function initialize_sbm_model(config::Config)
         indices_reverse,
         x_nc,
         y_nc,
-        dims_xy,
         nc,
         maxlayers = sbm.maxlayers,
     )