+++ src/sediment_model.jl
@@ -23,12 +23,6 @@ function initialize_sediment_model(config::Config)
     nc = NCDataset(static_path)
     dims = dimnames(nc[param(config, "input.subcatchment")])
 
-    # There is no need to permute the dimensions of the data, since the active indices are
-    # correctly calculated in both ways.
-    # The dimension order only needs to be known for interpreting the LDD directions
-    # and creating the coordinate maps.
-    dims_xy = dims[2] in ("y", "lat")
-
     subcatch_2d = ncread(nc, param(config, "input.subcatchment"); allow_missing = true)
     # indices based on catchment
     inds, rev_inds = active_indices(subcatch_2d, missing)
@@ -52,13 +46,9 @@ function initialize_sediment_model(config::Config)
     lake = ()
 
     # read x, y coordinates and calculate cell length [m]
-    y_nc = "y" in keys(nc.dim) ? ncread(nc, "y") : ncread(nc, "lat")
-    x_nc = "x" in keys(nc.dim) ? ncread(nc, "x") : ncread(nc, "lon")
-    if dims_xy
-        y = permutedims(repeat(y_nc, outer = (1, length(x_nc))))[inds]
-    else
-        y = repeat(y_nc, outer = (1, length(x_nc)))[inds]
-    end
+    y_nc = read_y_axis(nc)
+    x_nc = read_x_axis(nc)
+    y = permutedims(repeat(y_nc, outer = (1, length(x_nc))))[inds]
     cellength = abs(mean(diff(x_nc)))
 
     xl = fill(mv, n)
@@ -82,8 +72,6 @@ function initialize_sediment_model(config::Config)
     ols = OverlandFlowSediment{Float}(
         n = n,
         rivcell = rivcell,
-        h_riv = fill(mv, n),
-        q_riv = fill(mv, n),
         soilloss = fill(mv, n),
         erosclay = fill(mv, n),
         erossilt = fill(mv, n),
@@ -110,7 +98,6 @@ function initialize_sediment_model(config::Config)
         inlandlagg = fill(mv, n),
     )
 
-    pcr_dir = dims_xy ? permute_indices(pcrdir) : pcrdir
     graph = flowgraph(ldd, inds, pcr_dir)
 
     # River processes
@@ -136,7 +123,7 @@ function initialize_sediment_model(config::Config)
         lake = isempty(lake) ? nothing : lake.reverse_indices,
     )
     writer =
-        prepare_writer(config, reader, modelmap, indices_reverse, x_nc, y_nc, dims_xy, nc)
+        prepare_writer(config, reader, modelmap, indices_reverse, x_nc, y_nc, nc)
     close(nc)
 
     # for each domain save the directed acyclic graph, the traversion order,
@@ -210,12 +197,7 @@ function update(model::Model{N,L,V,R,W}) where {N,L,V<:LandSediment,R,W}
     do_river = get(config.model, "runrivermodel", false)::Bool
 
     if do_river
-        # Forcing come from lateral.land instead of netcdf directly
-        # Fix update_forcing in io to allow forcing for river (problem wih sel)
         inds_riv = network.index_river
-        lateral.river.h_riv .= lateral.land.h_riv[inds_riv]
-        lateral.river.q_riv .= lateral.land.q_riv[inds_riv]
-
         lateral.river.inlandclay .= lateral.land.inlandclay[inds_riv]
         lateral.river.inlandsilt .= lateral.land.inlandsilt[inds_riv]
         lateral.river.inlandsand .= lateral.land.inlandsand[inds_riv]