+++ test/runtests.jl
@@ -1,5 +1,6 @@
 ## load test dependencies and set paths to testing data
 using Dates
+using Downloads
 using LightGraphs
 using NCDatasets
 using StaticArrays
@@ -27,7 +28,7 @@ function testdata(version, source_filename, target_filename)
     target_path = joinpath(datadir, target_filename)
     base_url = "https://github.com/visr/wflow-artifacts/releases/download"
     url = string(base_url, '/', string('v', version), '/', source_filename)
-    isfile(target_path) || download(url, target_path)
+    isfile(target_path) || Downloads.download(url, target_path)
     return target_path
 end
 
@@ -37,16 +38,16 @@ staticmaps_lahn_path = testdata(v"0.2.1", "staticmaps-lahn.nc", "staticmaps-lahn
 forcing_moselle_path = testdata(v"0.2", "forcing-2000.nc", "forcing-moselle.nc")
 forcing_lahn_path = testdata(v"0.2", "forcing-lahn.nc", "forcing-lahn.nc")
 forcing_moselle_sed_path =
-    testdata(v"0.2", "forcing-moselle-sed.nc", "forcing-moselle-sed.nc")
+    testdata(v"0.2.3", "forcing-moselle-sed.nc", "forcing-moselle-sed.nc")
 staticmaps_moselle_sed_path =
-    testdata(v"0.2.2", "staticmaps-moselle-sed.nc", "staticmaps-moselle-sed.nc")
+    testdata(v"0.2.3", "staticmaps-moselle-sed.nc", "staticmaps-moselle-sed.nc")
 instates_moselle_sed_path =
     testdata(v"0.2", "instates-moselle-sed.nc", "instates-moselle-sed.nc")
 instates_moselle_path = testdata(v"0.2.1", "instates-moselle.nc", "instates-moselle.nc")
 forcing_sbm_gw_path =
     testdata(v"0.2.1", "forcing-sbm-groundwater.nc", "forcing-sbm-groundwater.nc")
 staticmaps_sbm_gw_path =
-    testdata(v"0.2.1", "staticmaps-sbm-groundwater.nc", "staticmaps-sbm-groundwater.nc")
+    testdata(v"0.2.2", "staticmaps-sbm-groundwater.nc", "staticmaps-sbm-groundwater.nc")
 lake_sh_1_path = testdata(v"0.2.1", "lake_sh_1.csv", "lake_sh_1.csv")
 lake_sh_2_path = testdata(v"0.2.1", "lake_sh_2.csv", "lake_sh_2.csv")
 lake_hq_2_path = testdata(v"0.2.1", "lake_hq_2.csv", "lake_hq_2.csv")